Ensure that wptserve shuts down subprocesses cleanly
We were calling terminate() on the server processes, which meant that
they didn't have a chance to shutdown properly and in particular could
end up failing to unlock cross-process shared locks.

This replaces the kill() method with a stop() method that arranges for
a clean shutdown.

It also removed the block argument from the Daemon process start
argument, since running the servers in a way that directly blocks the
main loop isn't really supported. Instead we always arrange for the
server to be started in a thread so we're later able to call
shutdown() on it.

There was also some signal handling logic that wasn't really doing
much, so that's replaced with just trapping the default exceptions.