Don't miss list item markers inside multicol containers.
With list-style-position:inside, we'd fail to find the marker if it was
inside a multicol container, because we'd look for it in the flow
thread, and give up, since the flow thread isn't the list-item (the
multicol container (the parent LayoutObject) is).

Flow threads should be ignored in LayoutNG. Use
GetLayoutObjectForParentNode() if we need to walk up the tree, to get
this right. Templatize GetLayoutObjectForParentNode(), so that it can
provide both a const and non-const version. CollectInlinesInternal()
works on non-const objects, and there doesn't seem to be an easy way
around that.

This fixes the final failures in multicol-span-all-list-item-001.html
and multicol-span-all-list-item-002.html, but since those tests were
quite complicated (and mostly about spanners, even if this CL has
nothing to do with spanners), I added a simpler test for this specific
issue.

Bug: 829028, 1130451
Change-Id: I6bcf10593a68c40cbc9adc11f9572c9aee580623