Correct list-item marker positioning in multicols.
Place list-item markers of a multicol container beside a spanner, if the
spanner has a baseline and there's no preceding first column with a
baseline. If a spanner has no baseline, also keep looking for suitable
columns after the spanner. If no suitable multicol child fragment has a
baseline, fall back to placing the marker at the beginning of the
multicol container. This matches the behavior in the block layout
algorithm.

In the simple cases we were already handling multicol list-item marker
positioning correctly, given that the first column had a baseline. In
such cases we'd lay out and add the marker inside the first column
fragment. Since the first column may not be the destination for the
marker, this CL will instead consistently put the marker as a direct
child of the multicol container (not inside a column). The block layout
algorithm will now always propagate the marker to the multicol container
if it's a list-item (rather than attempting to handle it on its own).
Remove two DCHECKs that are no longer sufficient. We could add list-item
markers as another exception to those DCHECKs, but removing them seems
better at this point.

Bug: 1130451
Change-Id: Idb4b950079d3cfdd7bc3a935b5b373a5a9f54aa1