Populate pointerId, pointerType for contextmenu event
For contextmenu as a pointer event pointerId and pointerType
were not populated.

In this CL we are populating pointerId/pointerType for contextmenu
coming from touch and mouse.
For touch we use a similar approach to crrev.com/c/2800231
for populating pointerId,pointerType for clicks generated in
GestureManager::HandleGestureTap.
The main change will be in EventHandler::SendContextMenuEvent
where we'll be passing id and pointer type to the call to
MouseEventManager::DispatchMouseEvent. This in turn will
populate pointerId and pointerType correctly when the contextmenu
PointerEvent is created.

Bug: 1150442
TEST: external/wpt/pointerevents/pointerevent_contextmenu_is_a_pointerevent.html
Change-Id: If2408d9cf0d5f00b08efcf2236ff7933472972ce