Populate pointerId, pointerType for clicks from touch
Pointer id and pointer type are not available in GestureManager
where we initiate the click from GestureTap events.

Pointer id algorithm is contained in PointerEventFactory. The problem
with surfacing pointer id in GestureManager is made harder because,
when we need to dispatch the click event, the pointer id for the
pointer event sequence that generates the click has already been
deleted from PointerEventFactory.

We will keep track of the last pointerdown event's pointer id, per
pointer type in PointerEventManager.

Using this approach, when we dispatch click event in GestureManager
we can then use PointerEventManager to access
the already deleted pointer id for the last pointer event sequence
that generated the click event. This works because we store the
pointer id of the last pointerdown event.

Bug: 1150593
TEST: third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_click_is_a_pointerevent.html
Change-Id: I7e921934f99c78298bad96b69aaa163e63c4be29