[GridNG] Fix invalidation logic for out of flow items via simplified layout
This CL addresses a failure where an absolutely positioned item changed
its position dynamically. Previously, a new |containing_block_rect| was
not being added nor re-computed in |SimplifiedLayoutAlgorithm| for
out of flow items that are children of the grid, thus their position was
incorrect. This change computes the new containing block by calling
existing helpers in |NGGridLayoutAlgorithm|. As a result, some methods
had to be changed and constructors overloaded in order to use what was
already computed and stored in |NGGridData|.

Bug: 1045599
Change-Id: Ic299a57aa5e69063c6f8d0f39c92772503b15f30